+++ backend/src/main/java/com/xenogames/common/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
-package com.gamevault.common;
+package com.xenogames.common;
 
-import com.gamevault.common.exception.BusinessException;
+import com.xenogames.common.exception.BusinessException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;