+++ backend/src/main/java/com/xenogames/common/Result.java
@@ -1,4 +1,4 @@
-package com.gamevault.common;
+package com.xenogames.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;