+++ backend/src/main/java/com/xenogames/config/WebMvcConfig.java
@@ -1,4 +1,4 @@
-package com.gamevault.config;
+package com.xenogames.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;